migrate names from lifecycle module to new class
